pyproject.toml
[project]
name = "leader-words-test-paper"
version = "1.0.0"
description = "A tool for generating word test papers with randomized questions"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.0.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.uv.workspace]
# 告訴 uv 這是一個簡單的腳本專案，不需要打包
members = []

[tool.black]
line-length = 88
target-version = ['py311']

